[YWA-1025] Add time off validity date change when its top changes

+++ app/services/employee_balance/manage_employee_balance_removal.rb
@@ -2,14 +2,14 @@ class ManageEmployeeBalanceRemoval
   attr_reader :new_date, :resource, :current_date
 
   def initialize(new_date, resource, current_date = resource.validity_date)
-    @new_date = new_date.try(:to_date)
+    @new_date = new_date
     @resource = resource
     @current_date = current_date
   end
 
   def call
     return unless !resource.time_off_policy.counter? && validity_date_changed?
-    new_date.blank? || moved_to_future? ? unassign_from_removal : create_or_assign_to_new_removal
+    new_date.blank? ? unassign_from_removal : create_or_assign_to_new_removal
   end
 
   private
@@ -17,11 +17,8 @@ class ManageEmployeeBalanceRemoval
   def unassign_from_removal
     resource_removal = resource.balance_credit_removal
     return unless resource_removal
-    if resource_removal.balance_credit_additions.count > 1
-      resource.update!(balance_credit_removal_id: nil)
-    else
-      resource_removal.destroy!
-    end
+    resource.update!(balance_credit_removal_id: nil)
+    resource_removal.destroy! if resource_removal.reload.balance_credit_additions.blank?
   end
 
   def create_or_assign_to_new_removal
@@ -36,7 +33,6 @@ class ManageEmployeeBalanceRemoval
     new_removal =
       Employee::Balance
       .removal_at_date(resource.employee_id, resource.time_off_category_id, new_date)
-
     new_removal.first_or_create do |removal|
       removal.effective_at = new_date
     end
